refactor: 피드백 반영
- domain 패키지가 application 패키지에 의존하는 문제 개선
- 인수 테스트 시나리오 추가
- 테스트 검증 메서드 추출
