+++ src/main/java/nextstep/subway/applicaion/PathService.java
@@ -5,7 +5,7 @@ import nextstep.subway.domain.Line;
 import nextstep.subway.domain.Path;
 import nextstep.subway.domain.PathType;
 import nextstep.subway.domain.Station;
-import nextstep.subway.domain.SubwayMap;
+import nextstep.subway.domain.service.SubwayMap;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
